# openapi.yaml
openapi: 3.1.0
info:
  title: AEGIS Scanner
  version: 1.3.0
  description: Semgrep-as-a-service for the AEGIS agent fleet. Guild imports this spec as a custom integration.
servers:
  - url: https://REPLACE-WITH-TUNNEL.trycloudflare.com
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-AEGIS-Key
  schemas:
    Finding:
      type: object
      required: [rule_id, path, start_line, end_line, lines, message, severity, fingerprint]
      properties:
        rule_id: { type: string }
        path: { type: string }
        start_line: { type: integer }
        end_line: { type: integer }
        lines: { type: string, description: matched source lines }
        message: { type: string }
        severity: { type: string, enum: [ERROR, WARNING, INFO] }
        cwe: { type: string }
        fingerprint: { type: string, description: "stable id = sha1(rule_id, path, lines) truncated to 12 chars" }
        fix: { type: string, description: replacement code suggested by the rule, when available }
        fix_hint: { type: string, description: how to fix when no mechanical fix exists }
        seen_before: { type: integer, description: times this fingerprint was seen across the fleet (ClickHouse) }
        dismissed_before: { type: boolean, description: triage already marked this fingerprint a false positive }
        repo_mttr_h: { type: number, description: mean time to remediate in this repo, hours }
        taint_rule_id: { type: string, description: "taint rule that also matched this line (its trace is attached) when a syntactic rule's id/fingerprint was kept" }
        trace_text: { type: string, description: "Semgrep's proven dataflow (--dataflow-traces) as one sentence: source -> propagators -> sink with file:line; only for taint rules. Quote it, do not infer the flow yourself." }
        dataflow_trace:
          type: array
          description: "steps of Semgrep's taint trace in order; only for taint rules"
          items:
            type: object
            required: [kind, path, line, code]
            properties:
              kind: { type: string, enum: [source, propagator, sink] }
              path: { type: string }
              line: { type: integer }
              code: { type: string }
...
    ScanResult:
      type: object
      required: [repo, sha, verdict, findings, ms]
      properties:
        repo: { type: string }
        sha: { type: string }
        base_sha: { type: string }
        verdict: { type: string, enum: [safe, unsafe] }
        findings:
          type: array
          items: { $ref: '#/components/schemas/Finding' }
        n_files: { type: integer }
        ms: { type: integer }
...
security:
  - apiKey: []
paths:
  /scan/diff:
    post:
      operationId: scan_diff
      summary: Scan only what a push introduced (Semgrep --baseline-commit). Answers "did this change make the repo unsafe?"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repo, base_sha, head_sha, agent]
              properties:
                repo: { type: string, description: owner/name }
                base_sha: { type: string, description: commit before the push }
                head_sha: { type: string, description: commit after the push }
                agent: { type: string, description: calling agent name, e.g. aegis-sentinel-01 }
      responses:
        '200':
          description: scan result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScanResult' }
  /scan/full:
    post:
      operationId: scan_full
      summary: Scan every tracked file of a repo at a commit.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repo, sha, agent]
              properties:
                repo: { type: string }
                sha: { type: string }
                agent: { type: string }
      responses:
        '200':
          description: scan result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScanResult' }
  /actions:
    post:
      operationId: record_action
      summary: Agent reports an action it took (stored in ClickHouse).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agent, repo, kind, ref]
              properties:
                agent: { type: string }
                repo: { type: string }
                kind: { type: string, enum: [status_set, issue_opened, issue_closed, pr_opened, pr_reviewed, dismissed, denied, email, verified, verify_failed, handoff_ok, handoff_rejected] }
                ref: { type: string, description: issue/PR number, sha, or url }
                fingerprint: { type: string }
                latency_ms: { type: integer }
                session_url: { type: string }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties: { ok: { type: boolean } }
...
  /insights:
    get:
      operationId: fleet_insights
      summary: Fleet-wide analytics from ClickHouse for the warden agent.
      parameters:
        - name: hours
          in: query
          schema: { type: integer, default: 24 }
      responses:
        '200':
          description: insights
          content:
            application/json:
              schema:
                type: object
                properties:
                  rising_repos: { type: array, items: { type: object } }
                  noisy_rules: { type: array, items: { type: object } }
                  reopened: { type: array, items: { type: object } }
                  agent_latency: { type: array, items: { type: object } }
                  handoffs:
                    type: object
                    description: handoff guard activity in the window
                    properties:
                      ok: { type: integer }
                      rejected: { type: integer }
                      last: { type: array, items: { type: object }, description: "last 10: {ts, from, to, kind, ok, rule_ids, repo, ref}" }
...
  /second-opinion:
    post:
      operationId: second_opinion
      summary: Scan a code snippet with the local AEGIS rules AND with Semgrep's hosted MCP server (mcp.semgrep.ai, tool semgrep_scan), and return both plus the overlap by line. available=false with a reason when the hosted server is down or needs OAuth (set SEMGREP_MCP_TOKEN); the local result is always returned.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
                language: { type: string, default: python }
                filename: { type: string }
      responses:
        '200':
          description: local and hosted findings
          content:
            application/json:
              schema:
                type: object
                properties:
                  available: { type: boolean }
                  local: { type: array, items: { type: object } }
                  hosted: { type: object, properties: { available: { type: boolean }, reason: { type: string }, findings: { type: array, items: { type: object } } } }
                  overlap_lines: { type: array, items: { type: integer } }
                  only_local_lines: { type: array, items: { type: integer } }
                  only_hosted_lines: { type: array, items: { type: integer } }
                  ms: { type: integer }